@@ -1,6 +1,6 @@
 {
   "name": "song-number",
-  "version": "0.1.2",
+  "version": "0.1.3",
   "author": "Florian Chiș",
   "homepage": "https://github.com/Fl0r14n/song-number.git",
   "scripts": {
@@ -9,52 +9,68 @@
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "electron": "npx cap open @capacitor-community/electron"
   },
   "dependencies": {
-    "@angular/common": "~11.0.2",
-    "@angular/core": "~11.0.2",
-    "@angular/forms": "~11.0.2",
-    "@angular/platform-browser": "~11.0.2",
-    "@angular/platform-browser-dynamic": "~11.0.2",
-    "@angular/router": "~11.0.2",
-    "@capacitor/android": "^2.4.3",
-    "@capacitor/core": "2.4.3",
-    "@ionic/angular": "^5.5.1",
-    "@ionic/storage": "^2.3.1",
-    "@ionic/pwa-elements": "^3.0.1",
+    "@angular/common": "~12.1.1",
+    "@angular/core": "~12.1.1",
+    "@angular/forms": "~12.1.1",
+    "@angular/platform-browser": "~12.1.1",
+    "@angular/platform-browser-dynamic": "~12.1.1",
+    "@angular/router": "~12.1.1",
+    "@capacitor/app": "1.0.2",
+    "@capacitor/camera": "^1.0.3",
+    "@capacitor/core": "3.1.2",
+    "@capacitor/haptics": "1.0.2",
+    "@capacitor/keyboard": "1.0.2",
+    "@capacitor/splash-screen": "^1.1.0",
+    "@capacitor/status-bar": "1.0.2",
+    "@capacitor/storage": "^1.0.3",
+    "@ionic/angular": "^5.5.2",
+    "@ionic/pwa-elements": "^3.0.2",
     "@ngx-translate/core": "^13.0.0",
-    "@ngx-translate/http-loader": "^6.0.0",
-    "core-js": "^3.8.0",
     "rxjs": "~6.6.3",
-    "tslib": "^2.0.3",
+    "tslib": "^2.3.0",
     "ubuntu-fontface": "^0.1.13",
-    "zone.js": "~0.11.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.2",
-    "@angular/cli": "~11.0.2",
-    "@angular/compiler": "~11.0.2",
-    "@angular/compiler-cli": "~11.0.2",
-    "@angular/language-service": "~11.0.2",
-    "@capacitor/cli": "2.4.3",
-    "@ionic/angular-toolkit": "^3.0.0",
-    "@types/jasmine": "~3.6.2",
-    "@types/jasminewd2": "~2.0.8",
-    "@types/node": "^14.14.10",
-    "codelyzer": "^6.0.1",
+    "@angular-devkit/build-angular": "~12.1.1",
+    "@angular-eslint/builder": "~12.0.0",
+    "@angular-eslint/eslint-plugin": "~12.0.0",
+    "@angular-eslint/eslint-plugin-template": "~12.0.0",
+    "@angular-eslint/template-parser": "~12.0.0",
+    "@angular/cli": "~12.1.1",
+    "@angular/compiler": "~12.1.1",
+    "@angular/compiler-cli": "~12.1.1",
+    "@angular/language-service": "~12.0.1",
+    "@capacitor-community/electron": "^3.0.2",
+    "@capacitor/android": "^3.1.2",
+    "@capacitor/cli": "3.1.2",
+    "@capacitor/ios": "^3.1.2",
+    "@ionic/angular-toolkit": "^4.0.0",
+    "@types/jasmine": "~3.6.0",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^12.11.1",
+    "@typescript-eslint/eslint-plugin": "4.16.1",
+    "@typescript-eslint/parser": "4.16.1",
     "cordova-chromecast": "git+https://github.com/jellyfin/cordova-plugin-chromecast.git",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~6.0.0",
-    "karma": "~5.2.3",
+    "eslint": "^7.6.0",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jsdoc": "30.7.6",
+    "eslint-plugin-prefer-arrow": "1.2.2",
+    "jasmine-core": "~3.8.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.3.2",
     "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
     "karma-coverage-istanbul-reporter": "~3.0.3",
-    "karma-jasmine": "~4.0.1",
-    "karma-jasmine-html-reporter": "^1.5.4",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
-    "ts-node": "~9.0.0",
-    "tslint": "~6.1.3",
-    "typescript": "~4.0.5"
+    "ts-node": "~8.3.0",
+    "typescript": "~4.2.4"
   },
   "description": "A small mobile app that will allow user to set a song number from a list of song books and cast it."
 }
