@@ -1,12 +1,27 @@
 import {NgModule} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
-import {RouteReuseStrategy, RouterModule, Routes} from '@angular/router';
+import {PreloadAllModules, RouteReuseStrategy, RouterModule, Routes} from '@angular/router';
 import {IonicModule, IonicRouteStrategy} from '@ionic/angular';
 import {AppComponent} from './app.component';
 import {TranslateLoader, TranslateModule} from '@ngx-translate/core';
-import {HttpClient, HttpClientModule} from '@angular/common/http';
-import {TranslateHttpLoader} from '@ngx-translate/http-loader';
-import {IonicStorageModule} from '@ionic/storage';
+import {HttpClientModule} from '@angular/common/http';
+import {Observable} from 'rxjs';
+import {en} from './i18n/en';
+import {ro} from './i18n/ro';
+
+export class I18nLoader implements TranslateLoader {
+
+  lang = {
+    en, ro
+  };
+
+  getTranslation(lang: string): Observable<any> {
+    return new Observable<any>(subscriber => {
+      subscriber.next(this.lang[lang]);
+      subscriber.complete();
+    });
+  }
+}
 
 const routes: Routes = [
   {
@@ -16,25 +31,20 @@ const routes: Routes = [
 ];
 
 @NgModule({
-  declarations: [AppComponent],
-  entryComponents: [],
   imports: [
     BrowserModule,
     IonicModule.forRoot(),
-    RouterModule.forRoot(routes),
-    IonicStorageModule.forRoot(),
+    RouterModule.forRoot(routes, {preloadingStrategy: PreloadAllModules}),
     HttpClientModule,
     TranslateModule.forRoot({
       loader: {
         provide: TranslateLoader,
-        useFactory: (http: HttpClient) => new TranslateHttpLoader(http, './assets/i18n/', '.json'),
-        deps: [HttpClient]
+        useClass: I18nLoader
       }
-    })
+    }),
   ],
-  providers: [
-    {provide: RouteReuseStrategy, useClass: IonicRouteStrategy}
-  ],
+  declarations: [AppComponent],
+  providers: [{provide: RouteReuseStrategy, useClass: IonicRouteStrategy}],
   bootstrap: [AppComponent]
 })
 export class AppModule {
