@@ -1,14 +1,12 @@
 import {Injectable} from '@angular/core';
-import {LoggerService} from './logger.service';
 import {from, Observable} from 'rxjs';
-import {CameraOptions, CameraResultType, CameraSource, Plugins} from '@capacitor/core';
 import {map} from 'rxjs/operators';
+import {Camera, CameraResultType, CameraSource, ImageOptions} from '@capacitor/camera';
 
-const {Camera} = Plugins;
 const width = 600;
 const height = 600;
 
-const cameraOptions = (source: CameraSource): CameraOptions => {
+const cameraOptions = (source: CameraSource): ImageOptions => {
   return {
     quality: 50,
     allowEditing: true,
@@ -25,10 +23,7 @@ const cameraOptions = (source: CameraSource): CameraOptions => {
 })
 export class CameraService {
 
-  constructor(private log: LoggerService) {
-  }
-
-  getPicture(source: CameraSource): Observable<string> {
+  getPicture$(source: CameraSource): Observable<string> {
     return from(Camera.getPhoto(cameraOptions(source))).pipe(
       map(img => `data:image/jpeg;base64,${img.base64String}`)
     );
